mobsf_cli_python/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "mobsf-cli"
version = "0.1.0"
description = "CLI wrapper for Mobile Security Framework (MobSF) REST API"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Wojciech Zurek", email = "mail@wojciechzurek.eu"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.28.0",
    "tabulate>=0.9.0",
    "python-dateutil>=2.8.0",
]

[project.scripts]
mobsf-cli = "cli.main:main"
